add usePool and useDailyTVL
