+++ tinlake-ui/utils/useDailyTVL.ts
@@ -1,0 +1,6 @@
+import { useQuery } from 'react-query'
+import Apollo from '../services/apollo'
+
+export function useDailyTVL() {
+  return useQuery('dailyTVL', () => Apollo.getPoolsDailyData(), { staleTime: 24 * 60 * 60 * 1000 })
+}
